refactor(similarity): extract BookSimilarityVectorFusion from embedding service
Move EMBEDDING_DIMENSION, fuseSectionEmbeddings, and both vectorNorm
overloads into a dedicated BookSimilarityVectorFusion @Component so the
service stays focused on scheduling, dedup, and orchestration while the
pure vector math is owned by one class driven by the active fusion
policy weights.

Service drops from 390 to 346 lines, below the 350 [LOC1a] ceiling.

Corrected following review from Greptile.
